set services types on a generic folder
