--- src/services/geo-processing-services/contextual/types.d.ts
@@ -1,15 +1,0 @@
-export interface GetCrfDataProps {
-  aoiFeatureGeometry: {
-    rings: number[][];
-  };
-  rings: number[][];
-  toJSON: () => Record<string, unknown>;
-}
-
-export interface JobInfoProps {
-  jobId: number;
-  fetchResultData: (outputParamKey: string) => function;
-  waitForJobCompletion: (jobId: number) => function;
-  catch: (error: Error) => unknown;
-  then: () => unknown;
-}
